backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "firebase-admin>=7.0.0",
    "google-adk>=0.0.1",
    "google-generativeai>=0.8.5",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "streamlit>=1.47.0",
    "tavily-python>=0.7.10",
    "uvicorn>=0.35.0",
]
